+++ lib/src/matomo.dart
@@ -312,9 +312,9 @@ class MatomoTracker {
     assert(initialized);
     log.finest('Processing queue ${_queue.length}');
 
-    if (_queue.isEmpty) {
-      _dispatcher.sendPing(this);
-    }
+    // if (_queue.isEmpty) {
+    //   _dispatcher.sendPing(this);
+    // }
 
     while (_queue.isNotEmpty) {
       final event = _queue.removeFirst();